services/auth/requirements.txt
psycopg2==2.9.10
asyncpg==0.30.0
pyJWT==2.10.0
fastapi==0.115.5
orjson==3.10.5
pydantic_settings==2.6.1
redis==5.2.0
alembic==1.14.0
SQLAlchemy==2.0.36
werkzeug==3.1.3
uvicorn==0.32.1